Attempt to hook up dependencies
