+++ package.json
@@ -70,7 +70,6 @@
   ],
   "dependencies": {
     "@bessonovs/node-http-router": "^0.0.5",
-    "@koa/router": "^10.0.0",
     "@rdfjs/data-model": "^1.2.0",
     "@types/arrayify-stream": "^1.0.0",
     "@types/async-lock": "^1.1.2",
@@ -92,11 +91,6 @@
     "express": "^4.17.1",
     "fetch-sparql-endpoint": "^1.8.0",
     "handlebars": "^4.7.6",
-    "koa": "^2.13.0",
-    "koa-body": "^4.2.0",
-    "koa-ejs": "^4.3.0",
-    "koa-mount": "^4.0.0",
-    "koa-router": "^10.0.0",
     "lodash": "^4.17.20",
     "mime-types": "^2.1.27",
     "n3": "^1.6.4",
@@ -117,11 +111,6 @@
   "devDependencies": {
     "@microsoft/tsdoc-config": "^0.13.6",
     "@types/jest": "^26.0.13",
-    "@types/koa": "^2.11.6",
-    "@types/koa__router": "^8.0.3",
-    "@types/koa-ejs": "^4.2.2",
-    "@types/koa-mount": "^4.0.0",
-    "@types/koa-router": "^7.4.1",
     "@types/rimraf": "^3.0.0",
     "@types/supertest": "^2.0.10",
     "@typescript-eslint/eslint-plugin": "^4.1.1",
