+++ src/identity/email-password/EmailPasswordIdentityProviderHandler.ts
@@ -1,37 +1,28 @@
 import type { IncomingMessage, ServerResponse } from 'http';
-import type Koa from 'koa';
 import { getLoggerFor } from '../../logging/LogUtil';
 import { HttpHandler } from '../../server/HttpHandler';
 import type { Guarded } from '../../util/GuardedStream';
-import getApp from './tempServer';
+import type { OidcProviderFactory } from '../provider/OidcProviderFactory';
 
 interface EmailPasswordIdentityProviderHandlerArgs {
-  optional?: string;
+  oidcProviderFactory: OidcProviderFactory;
 }
 
 export class EmailPasswordIdentityProviderHandler extends HttpHandler {
-  private readonly app: Koa;
+  private readonly oidcProviderFactory: OidcProviderFactory;
   private readonly logger = getLoggerFor(this);
 
   public constructor(args: EmailPasswordIdentityProviderHandlerArgs) {
     super();
-    Object.assign(this, args);
-    this.app = getApp();
-    // TODO [>1.0.0] Reenable error passthrough
-    // this.app.onerror = (err): void => {
-    //   throw err;
-    // };
+    this.oidcProviderFactory = args.oidcProviderFactory;
   }
 
   public async handle(input: {
     request: Guarded<IncomingMessage>;
     response: ServerResponse;
   }): Promise<void> {
-    this.logger.info('Handling');
-    const callbackPromise = (this.app.callback()(
-      input.request,
-      input.response,
-    ) as unknown) as Promise<void>;
-    await callbackPromise;
+    const provider = await this.oidcProviderFactory.createOidcProvider();
+    const thing = provider.callback(input.request, input.response);
+    this.logger.info(thing as unknown as string);
   }
 }
